feat(cli): refactor the service
