+++ packages/cli/src/util/time.ts
@@ -1,0 +1,4 @@
+export function sleep(ms: number): Promise<void> {
+    if (ms <= 0) return Promise.resolve();
+    return new Promise((resolve) => setTimeout(resolve, ms));
+}
